Added reindexing functionality and options to TimeseriesRefCriterion.cmpare.
The new functionality allows one to specify join reindexing of the reference
and input data to be applied before the comparison operation is carried out.
The IAM COMPACT timeseries criterion is adjusted so that its `compare` method
performs an inner join (after having broadcast the reference data), which means
that the result only contains combinations of model, scenario, region and
variable that occur in the input data, rather than returning a lot of all-NA
rows for regions that are present in the reference data but not in the input
data.