Preliminary commit of updated notebook script.
1st modelling cycle notebook .py file that contains working code to use the
new TimeseriesRefComparisonAndTargetOutput class to output GDP and population
harmonization vetting.

This is just the code, and just that one .py file. Markdown explanations still
need to be added and adjusted, the file needs to be exported to .ipynb format,
and the general files (not specific to the 1st modelling cycle) also need to be
updated.