Removed requirements that target be not-None in TimeSeriesRefComparisonAndTargetOutput.
Done to allow for passing in CriterionTargetRange subclass types where default
values for `target` and `range` have already been set. Consequently also
changed the method signature of the `target` parameter in
`CriterionTargetRange.__init__` and added a check in the __init__ method that
target has in fact been set to a float (otherwise raising an error that says
`target` must either be specified or the caller must use a subclass that
implements a default value).