Tracking D4.3 GDP/population Excel file directly.
Tracking the file with git instead of DVC, since it is used in the published
Python package. This only applies to the IAMCformat file, which is the one that
is actually used. The other Excel files in different formats (e.g., with ISO3
country names) are still tracked with DVC here in the dev_main branch, but will
be removed from the published package entirely.